requirements.txt
pandas
streamlit
python-dotenv
openai
plotly
mistralai
streamlit-aggrid
